+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/components/ValuesList.kt
@@ -43,8 +43,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
 import br.com.rbrthmn.R
+import br.com.rbrthmn.ui.financialcompanion.uistates.FinancialOverviewUiState
 import br.com.rbrthmn.ui.financialcompanion.utils.valueWithCurrencyString
-import br.com.rbrthmn.ui.financialcompanion.viewmodels.FinancialOverviewUiState
 
 @Composable
 fun ValuesList(
