+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/HomeScreen.kt
@@ -44,6 +44,7 @@ import br.com.rbrthmn.ui.financialcompanion.components.LastMonthDifferenceCard
 import br.com.rbrthmn.ui.financialcompanion.components.MonthlyLimitCard
 import br.com.rbrthmn.ui.financialcompanion.components.MonthSelectionTopBar
 import br.com.rbrthmn.ui.financialcompanion.navigation.NavigationDestination
+import br.com.rbrthmn.ui.financialcompanion.uistates.FinancialOverviewUiState
 import br.com.rbrthmn.ui.financialcompanion.utils.MonthsOfTheYear
 import br.com.rbrthmn.ui.financialcompanion.viewmodels.FinancialOverviewUiState
