+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/uistates/FinancialOverviewUiState.kt
@@ -1,0 +1,32 @@
+/*
+ *
+ * Copyright (C) 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Modifications made by Roberto Kenzo Hamano, 2024
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package br.com.rbrthmn.ui.financialcompanion.uistates
+
+import br.com.rbrthmn.R
+
+data class FinancialOverviewUiState(
+    val name: String = "",
+    val value: String = "",
+    val balanceCurrency: String = "",
+    val financialInstitutionName: String = "",
+    val financialInstitutionIcon: Int = R.drawable.account_balance,
+    val canValueBeEdited: Boolean = true
+)
